requirements.txt
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
findspark==2.0.1
greenlet==3.0.3
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kafka-python==2.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
Pygments==2.18.0
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
six==1.16.0
spark==0.2.1
SQLAlchemy==2.0.30
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
wcwidth==0.2.13
